requirements.txt
google-adk==1.14.0
langchain-community==0.3.27
wikipedia==1.4.0
google-auth>=2.0.0
google-cloud-bigquery
toolbox-core>=0.3.0
python-dotenv
uvicorn[standard]>=0.30.0
pandas
numpy
